// provider_client/src/api/api.ts
export interface ServiceDto {
  id: number;
  providerId: number;
  title: string;
  photo_url: string[];
  duration: number;
  start_time: number;
  end_time: number;
  category: string;
  days_of_week: string[];
  description: string;
  price: number;
  status: 'active' | 'deleted' | 'stopped';
  /** @format date-time */
  createdAt?: string;
  /** @format date-time */
  updatedAt?: string;
}